requirements.txt
streamlit
langchain-google-genai
langchain-openai
google-search-results
requests
python-dotenv
parsera
googlesearch-python
playwright
gtts
setuptools
beautifulsoup4
langdetect